Enable tests
