+++ src/test/groovy/info/dennis_weber/unfima/api/RootSpecification.groovy
@@ -1,0 +1,14 @@
+package info.dennis_weber.unfima.api
+
+import info.dennis_weber.unfima.api.helpers.AbstractUnfimaSpecification
+
+class RootSpecification extends AbstractUnfimaSpecification {
+
+    def "root is reachable"() {
+        when:
+        client.get("/")
+
+        then:
+        client.response.status.code == 200
+    }
+}
